+++ src/index.ts
@@ -1,0 +1,5 @@
+import APIv1 from "./data/google-fonts-v1.json";
+import APIv2 from "./data/google-fonts-v2.json";
+import APIVariable from "./data/variable.json";
+
+export { APIv1, APIv2, APIVariable };